Remove server user config and fix tests.
